# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ferrix API (Ajira Copilot)",
    "version": "1.0.0",
    "summary": "Voice-first worker intake, explainable matching, and the human trust pipeline — as a public API.",
    "description": "Ferrix (Ajira Copilot) public API. Auth contract: mutating routes require 'Authorization: Bearer $FERRIX_API_TOKEN' (or 'x-api-key') when that env var is set; unset = open judge demo. EVERY response carries the 'x-ferrix-auth' header (open-demo | enforced). Hand-maintained; kept in lockstep with docs/API.md and the QA suites.",
    "contact": {
      "name": "Ferrix (Ajira Copilot)",
      "url": "https://github.com/"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev server"
    }
  ],
...
  "paths": {
...
    "/api/analyze": {
      "post": {
        "operationId": "postAnalyze",
        "tags": [
          "Analyze"
        ],
        "summary": "Structure the profile, score matches, open/refresh the case (guarded mutation)",
        "description": "Runs the transcript of `candidateId` through the LLM to produce a Zod-shaped `StructuredProfile` (scripted persona fallback if the LLM fails), persists it, then scores active opportunities with the deterministic explainable matcher.\n\nEvery persisted `Match` row is stamped with `matcherVersion`, and the response echoes it (top level + per match). `DELETE`-and-recreate semantics: prior matches for the candidate are replaced. Reuses the candidate's open (non-resolved) case instead of minting a duplicate; creates a case (`AJR-…` ref) with `structured` events otherwise.\n\n`mode` is `live` when the LLM produced the profile, `fallback` when the persona baseline was used.\n\nRate limit: 10 requests/min per IP (LLM-heavy; 429 with `Retry-After`).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Structured profile, ranked explainable matches (max 4), and the target case.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON or body failing the Zod contract (`candidateId` required).",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Candidate not found.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Per-IP rate limit exceeded (10/min).",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the window resets.",
                "schema": {
                  "type": "string"
                }
              },
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/asset": {
      "post": {
        "operationId": "postAsset",
        "tags": [
          "Assets"
        ],
        "summary": "Generate a WhatsApp catalog or plain-text CV (guarded mutation)",
        "description": "Uses the candidate's persisted structured profile (so `/api/analyze` must have run first, else `400 \"Profile the candidate first\"`) to generate either a `catalog` (WhatsApp-ready stock list) or a `cv` (plain-text CV). Falls back to a deterministic template (mode `fallback`) when the LLM is unavailable.\n\nThere is no `GET /api/asset` — assets are generated on demand, not stored.\n\nRate limit: 20 requests/min per IP (429 with `Retry-After`).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated text; `mode` is `generated` (LLM) or `fallback` (template).",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON, bad `kind`, or the candidate has no structured profile yet.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Candidate not found.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Per-IP rate limit exceeded (20/min).",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the window resets.",
                "schema": {
                  "type": "string"
                }
              },
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "FERRIX_API_TOKEN. Required on mutating endpoints only when the deployment sets FERRIX_API_TOKEN (auth mode `enforced`; watch the `x-ferrix-auth` response header). `x-api-key: <token>` is accepted as an equivalent alternative header. In open demo mode (`x-ferrix-auth: open-demo`) no token is needed."
      }
    },
    "headers": {
      "XFerrixAuth": {
        "description": "Auth mode of the deployment for this request: `open-demo` (mutations open) or `enforced` (mutations require the token). Contract: every response carries it — today it is verifiably present on success responses and 401 auth rejections; the shared 400/404/429 helpers are being wired to stamp it as well.",
        "schema": {
          "type": "string",
          "enum": [
            "open-demo",
            "enforced"
          ]
        }
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Auth is enforced and the caller did not present the coordinator token (`Authorization: Bearer <token>` or `x-api-key`).",
        "headers": {
          "x-ferrix-auth": {
            "$ref": "#/components/headers/XFerrixAuth"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServerError": {
        "description": "Unexpected server error.",
        "headers": {
          "x-ferrix-auth": {
            "$ref": "#/components/headers/XFerrixAuth"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Short machine-readable error (first Zod issue message, or a stable label)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. Present on 401 auth failures."
          }
        }
      },
...
      "AnalyzeRequest": {
        "type": "object",
        "required": [
          "candidateId"
        ],
        "properties": {
          "candidateId": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "StructuredProfile": {
        "type": "object",
        "required": [
          "name",
          "summary",
          "skills",
          "tags",
          "experience",
          "digitalLiteracy",
          "availability",
          "goal",
          "constraints",
          "confidence",
          "aiSummary",
          "verificationFlags"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 8
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Lowercase slug tags used by the matcher (e.g. \"customer-service\")."
          },
          "experience": {
            "type": "string"
          },
          "digitalLiteracy": {
            "type": "string"
          },
          "availability": {
            "type": "string"
          },
          "goal": {
            "type": "string"
          },
          "constraints": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "confidence": {
            "type": "number",
            "minimum": 0.3,
            "maximum": 0.99
          },
          "aiSummary": {
            "type": "string",
            "description": "AI narration; always recommends human verification before placement."
          },
          "verificationFlags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "OpportunityRef": {
        "type": "object",
        "required": [
          "id",
          "title",
          "type",
          "provider",
          "location",
          "payRange",
          "description"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "COURSE",
              "GIG",
              "JOB",
              "MICROWORK"
            ]
          },
          "provider": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "payRange": {
            "type": "string"
          },
          "duration": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": "string"
          }
        }
      },
      "AnalyzeMatch": {
        "type": "object",
        "required": [
          "matcherVersion",
          "id",
          "score",
          "reasons",
          "opportunity"
        ],
        "properties": {
          "matcherVersion": {
            "type": "string",
            "const": "v1",
            "description": "Scorer version that produced this score."
          },
          "id": {
            "type": "string",
            "description": "Opportunity id."
          },
          "score": {
            "type": "number",
            "minimum": 0.05,
            "maximum": 0.98
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Human-readable explanations (why this match)."
          },
          "opportunity": {
            "$ref": "#/components/schemas/OpportunityRef"
          }
        }
      },
      "AnalyzeCaseSummary": {
        "type": "object",
        "required": [
          "id",
          "ref",
          "priority",
          "status",
          "request"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "ref": {
            "type": "string",
            "description": "Human reference, e.g. AJR-1007."
          },
          "priority": {
            "type": "string",
            "enum": [
              "high",
              "normal",
              "low"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "accepted",
              "contacted",
              "placed",
              "resolved"
            ]
          },
          "request": {
            "type": [
              "string",
              "null"
            ],
            "description": "The candidate's stated goal."
          }
        }
      },
      "AnalyzeResponse": {
        "type": "object",
        "required": [
          "matcherVersion",
          "profile",
          "matches",
          "case",
          "mode"
        ],
        "properties": {
          "matcherVersion": {
            "type": "string",
            "const": "v1",
            "description": "Version of the matcher that produced these scores; also stamped on each persisted Match row."
          },
          "profile": {
            "$ref": "#/components/schemas/StructuredProfile"
          },
          "matches": {
            "type": "array",
            "maxItems": 4,
            "items": {
              "$ref": "#/components/schemas/AnalyzeMatch"
            }
          },
          "case": {
            "$ref": "#/components/schemas/AnalyzeCaseSummary"
          },
          "mode": {
            "type": "string",
            "enum": [
              "live",
              "fallback"
            ],
            "description": "How the profile was produced (LLM vs persona baseline)."
          }
        }
      },
...
      "AssetRequest": {
        "type": "object",
        "required": [
          "candidateId",
          "kind"
        ],
        "properties": {
          "candidateId": {
            "type": "string",
            "minLength": 1
          },
          "kind": {
            "type": "string",
            "enum": [
              "catalog",
              "cv"
            ]
          }
        }
      },
      "AssetResponse": {
        "type": "object",
        "required": [
          "text",
          "mode"
        ],
        "properties": {
          "text": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": [
              "generated",
              "fallback"
            ]
          }
        }
      }
    }
  }
}